@@ -1,28 +1,18 @@
 name: fryo
 description: A food ordering app template kit.
 
-# The following defines the version and build number for your application.
-# A version number is three numbers separated by dots, like 1.2.43
-# followed by an optional build number separated by a +.
-# Both the version and the builder number may be overridden in flutter
-# build by specifying --build-name and --build-number, respectively.
-# In Android, build-name is used as versionName while build-number used as versionCode.
-# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
-# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
-# Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 version: 1.0.0+1
-
 environment:
-  sdk: ">=2.4.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=3.7.8 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   page_transition: '^2.0.9'
-  smooth_star_rating: '^1.1.1'
-  firebase_core: '^2.8.0'
-  firebase_auth: '^4.3.0'
+  flutter_rating_bar: ^4.0.1
+  firebase_core: '^2.9.0'
+  firebase_auth: '^4.4.0'
   google_sign_in: '^6.0.2'
   http: '^0.13.5'
   cupertino_icons: '^1.0.5'
@@ -39,17 +29,13 @@ dev_dependencies:
   freezed: ^2.3.2
   json_serializable: ^6.6.1
 
-# For information on the generic Dart part of this file, see the
-# following page: https://www.dartlang.org/tools/pub/pubspec
-
-# The following section is specific to Flutter.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
   uses-material-design: true
 
+  # TODO(lamuguo): 本地图片只用于app的渲染，不应与product / store的定制化相关。下面的数字图像应该都要删掉。
   # To add assets to your application, add an assets section, like this:
   assets:
    - images/welcome.png
@@ -62,11 +48,6 @@ flutter:
    - images/7.png
    - images/8.png
    - images/9.png
-
-
-
-
-
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.io/assets-and-images/#resolution-aware.
@@ -106,7 +87,5 @@ flutter:
       fonts:
         - asset: fonts/Fryo.ttf
 
-
-        
   # For details regarding fonts from package dependencies,
   # see https://flutter.io/custom-fonts/#from-packages
