@@ -3,10 +3,11 @@ import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 
 import '../shared/config.dart';
+import 'api_client.dart';
 
 // TODO(lamuguo): Move the function to api_client.dart
 Future<void> backendLogin() async {
-  String token = await FirebaseAuth.instance.currentUser.getIdToken(true);
+  String token = await getCurrentToken();
   print('xfguo: id token = ${token}');
   final response = await http.get(
     Uri.parse("${Config.instance.apiUrl}/login"),
@@ -23,11 +24,11 @@ Future<void> backendLogin() async {
 }
 
 class UserProvider with ChangeNotifier {
-  User _user;
-  User get user => _user;
+  User? _user = null;
+  User? get user => _user;
   bool get isLoggedIn => _user != null;
 
-  void login(User user) {
+  void login(User? user) {
     backendLogin();
 
     _user = user;
@@ -36,7 +37,6 @@ class UserProvider with ChangeNotifier {
 
   Future<void> logout() async {
     await FirebaseAuth.instance.signOut();
-    _user = null;
     notifyListeners();
   }
 }
